feat: move images from public to src/assets for Storybook compatibility
- Moved all images from public/images/ to src/assets/images/
- Updated all component imports to use @/ alias for cleaner paths
- Added static directory configuration to Storybook main.ts
- Added TypeScript declarations for image file types in global.d.ts
- Updated components: AuthUI, Header, NodeDetails, NodeEditModal
- Removed duplicate image files from public folder

This enables Storybook to properly import and display images while maintaining
clean import paths using the @/ alias throughout the codebase.
