+++ src/assets/images/logo_yellow.png
