+++ src/global.d.ts
@@ -2,3 +2,18 @@ declare module '*.svg' {
     const content: string
     export default content
 }
+
+declare module '*.png' {
+    const content: string
+    export default content
+}
+
+declare module '*.jpg' {
+    const content: string
+    export default content
+}
+
+declare module '*.jpeg' {
+    const content: string
+    export default content
+}
